+++ changelog.md
@@ -22,3 +22,14 @@
 
 ### Notes
 - I am really unhappy with how the script calculates the percentages, since bash does not deal with floating points without some extra work put into it. I don't think this is a big deal for what we need this script, but if we decide that we need precision later on, it might be worth it rewriting the whole thing in python, or deal with actual review numbers rather than percentages.
+
+## 2025-06-14 (Roberto)
+- Started working on the summarization of the audience inputs by creating a script capable of summarizing the fake reviews by using the Ollama AI models. It can be found at review-summarization/summarizer.sh
+- Extracted some summaries and started working on prompting and experimenting with different models for that. Weirdly, deepseek was worse at following the prompt summarization rules than weaker models. This is not all bad, as the summarization needs to happen quicly in between movements. All summaries can be found at review-summarization/ under the name pattern *model used*-sum-*review tone*-reviews.txt
+- Improved fake-review/review-generator.sh, and generated the review databases again.
+- Added changelog.md link to README.md
+
+### Notes
+- Using the word "review" to describe it to the AI proved to be a big mistake so all the scripts and prompts had to be reworked to avoid that word and get more precise AI outputs. Because of that, I decided to redo all of the fake review databases.
+- Sometimes, prompted rules are broken by the AI when summarizing too many audience, since the prompt becomes bigger than the context window of the AI. Despite the context being a problem, adding more details to prompt seems to make it better for mistral so I am trying that. A possible hack hack could be to resumarize the first output of the summarize (creating the first output as a temp file). The problem with this hack is that would make using slower and more powerfull models not possible since it would take too long between movements.
+- Luckly, mistral:7b is extremelly fast summarizing. I will start testing other models for that and try to use word cloud as a tool to quicly, visualy compare summarizations of the same database.